Renamed sync summon space roles
