+++ background/cron.ts
@@ -7,7 +7,7 @@ import { task as archiveTask } from './tasks/deleteArchivedPages';
 import { task as processWebhookMessages } from './tasks/processWebhookMessages';
 import { refreshBountyApplications } from './tasks/refreshBountyApplications/task';
 import { task as notificationTask } from './tasks/sendNotifications';
-import { syncSummonSpaceRoles } from './tasks/syncSummonSpaceRoles/task';
+import { syncSummonSpacesRoles } from './tasks/syncSummonSpaceRoles/task';
 import { task as proposalTask } from './tasks/updateProposalStatus';
 import { task as voteTask } from './tasks/updateVotesStatus';
 import { task as verifyTokenGateMembershipsTask } from './tasks/verifyTokenGateMemberships';
@@ -39,7 +39,7 @@ cron.schedule('*/30 * * * *', refreshBountyApplications);
 cron.schedule('*/30 * * * *', countAllSpacesBlocksTask);
 
 // Sync summon space roles every day at midnight
-cron.schedule('0 0 * * *', syncSummonSpaceRoles);
+cron.schedule('0 0 * * *', syncSummonSpacesRoles);
 
 const port = process.env.PORT || 4000;
