+++ background/tasks/syncSummonSpaceRoles/task.ts
@@ -1,9 +1,9 @@
 import { log } from '@charmverse/core/log';
 import { prisma } from '@charmverse/core/prisma-client';
 
-import { syncSpaceRole } from 'lib/summon/syncSpaceRole';
+import { syncSummonSpaceRoles } from 'lib/summon/syncSummonSpaceRoles';
 
-export async function syncSummonSpaceRoles() {
+export async function syncSummonSpacesRoles() {
   const spaces = await prisma.space.findMany({
     select: {
       id: true
@@ -21,7 +21,7 @@ export async function syncSummonSpaceRoles() {
 
   for (const spaceId of spaceIds) {
     try {
-      await syncSpaceRole({ spaceId });
+      await syncSummonSpaceRoles({ spaceId });
     } catch (err: any) {
       log.error(`Error syncing space role for space ${spaceId}: ${err.stack || err.message || err}`, { err });
     }
